Update Dockerfile to fix volume permissions
When mounting a volume on a linux hosts, sometimes the mounted folder
maintains the uid and permissions of the host. In this case, the user
does not have permissions to modify files in the workdir, or create new
ones, like when running make. This commit changes the way the container
manages permissions, giving all files in $HOME group write
permissions for the group 'users'. Furthermore, on running, it fixes
the permissions of the mounted folder, to make it writable by group
'users'